wpa_s: enable mesh HT mode
Add a new option "mesh_ht_mode" that specifies the HT mode for the
mesh, with this option on, mesh beacons, actions frames, and probe
responses with include the appropriate HT information elements.

Change the mesh tests to check for the presence of HT in the scan
results, for this to work "disable_scan_offload" must be added to
p2p*.conf.

Signed-off-by: Ashok Nagarajan <ashok.dragon@gmail.com>
Signed-off-by: Javier Cardona <javier@cozybit.com>
[Original implementation by Chun-Yeow Yeoh <yeohchunyeow@gmail.com>]
Signed-hostap: Jason Mobarak <x@jason.mobarak.name>

---
Changes since v2:
 - fixup IE order [Yu Niiro]